Initial BSP for msp430g2553

+++ pfal/sdmmc/pff.h
@@ -42,8 +42,8 @@ typedef uint32_t		DWORD;	/* 32-bit unsigned integer */
 #else  	/* Earlier than C99 */
 typedef unsigned int	UINT;	/* int must be 16-bit or 32-bit */
 typedef unsigned char	BYTE;	/* char must be 8-bit */
-typedef unsigned short	WORD;	/* 16-bit unsigned integer */
-typedef unsigned short	WCHAR;	/* 16-bit unsigned integer */
+typedef unsigned long	WORD;	/* 16-bit unsigned integer */
+typedef unsigned long	WCHAR;	/* 16-bit unsigned integer */
 typedef unsigned long	DWORD;	/* 32-bit unsigned integer */
 #endif
 #define PF_INTDEF 1
